+++ R/fake-http.R
@@ -1,12 +1,12 @@
 #' Make all HTTP requests return a fake 'response' object
 #'
 #' In this context, HTTP verb functions raise a 'message' so that test code can
-#' assert that the requests are made. Unlike \code{\link{without_internet}},
+#' assert that the requests are made. Unlike [without_internet()],
 #' the HTTP functions do not error and halt execution, instead returning a
-#' \code{response}-class object so that code calling the HTTP functions can
+#' `response`-class object so that code calling the HTTP functions can
 #' proceed with its response handling logic and itself be tested.
 #' @param expr Code to run inside the fake context
-#' @return The result of \code{expr}
+#' @return The result of `expr`
 #' @examples
 #' with_fake_HTTP({
 #'     expect_GET(req1 <- httr::GET("http://example.com"), "http://example.com")